Allow more fine-grained control with tox (#41)
This adds a few quality-of-life improvements to our test runner. A user no longer has to install the test dependencies separately; this will be handled by tox.

Additionally, it is possible to pass command line arguments through tox to cwltool. The intended use for this is that users can now run the CWL validations on systems where neither NodeJS nor Docker are available if they have access to apptainer (--singularity), but other flags (such as --strict) could be passed as well.

Also, it removes python 3.9 from the python version the test suite will target.