Add CHANGELOG.md
After moving to Github we've lost easy access to the changes between
versions of the pipeline. I've placed them in a markdown file. I think
it's easier to keep track of future changes by referring to this file
instead of listing them in the tag/release description, since it makes
fewer assumptions about where the code is stored.